OM5K-8850 show error msg if tag field is blank and stays in box

+++ client/src/components/i-button/i-button.jsx
@@ -1,30 +1,45 @@
 import React, { useState } from 'react';
-import { Modal, Input } from 'antd';
+import { Modal, Input, message } from 'antd';
 
-const KaytxtInput = ({ setSearch, enterButton }) => {
+const KaytxtInput = ({ setSearch, enterButton, prompt }) => {
   const [data, setData] = useState('');
+  const [error, setError] = useState('');
 
   const onChange = (event) => {
     console.log(event.target.value);
     console.log(event.target.value.toUpperCase());
     setData(event.target.value.toUpperCase());
+    let card_txt = event.target.value.toUpperCase().trim();
+    if (card_txt.length === 0) {
+      setError(prompt);
+    } else {
+      setError('');
+    }
   };
 
   return (
-    <Input.Search
-      style={{ marginTop: 10 }}
-      enterButton={enterButton}
-      onChange={onChange}
-      value={data}
-      onSearch={(value) => {
-        let card_txt = value.trim();
-        card_txt = card_txt.substring(2, card_txt.length - 2);
-        setSearch(card_txt);
-        Modal.destroyAll();
-      }}
-      // onPressEnter={Modal.destroyAll()}
-      autoFocus={true}
-    />
+    <>
+      <Input.Search
+        style={{ marginTop: 10 }}
+        enterButton={enterButton}
+        onChange={onChange}
+        value={data}
+        onSearch={(value) => {
+          let card_txt = value.trim();
+          if (card_txt.length === 0) {
+            // message.error(prompt);
+            setError(prompt);
+            return;
+          }
+          card_txt = card_txt.substring(2, card_txt.length - 2);
+          setSearch(card_txt);
+          Modal.destroyAll();
+        }}
+        // onPressEnter={Modal.destroyAll()}
+        autoFocus={true}
+      />
+      {error.length > 0 && <div style={{ color: 'red' }}>{error}</div>}
+    </>
   );
 };
 
@@ -40,7 +55,11 @@ const IButton = ({ setSearch, t, buttonType }) => {
     content: (
       <>
         <span>{t('descriptions.iButtonLookUp')}</span>
-        <KaytxtInput setSearch={setSearch} enterButton={t('operations.' + buttonType)} />
+        <KaytxtInput
+          setSearch={setSearch}
+          enterButton={t('operations.' + buttonType)}
+          prompt={t('validate.set')}
+        />
         <span>&nbsp;&nbsp;</span>
       </>
     ),
